+++ examples/stackflow-spa/src/seed-design/ui/extended-action-sheet.tsx
@@ -3,12 +3,19 @@
 import { ExtendedActionSheet as SeedExtendedActionSheet } from "@seed-design/react";
 import { forwardRef } from "react";
 
-export interface ExtendedActionSheetRootProps extends SeedExtendedActionSheet.RootProps {}
+/**
+ * @deprecated Use menu-sheet instead.
+ */
+export interface ExtendedActionSheetRootProps
+  extends SeedExtendedActionSheet.RootProps {}
 
 /**
  * @see https://seed-design.io/react/components/extended-action-sheet
+ * @deprecated Use menu-sheet instead.
  */
-export const ExtendedActionSheetRoot = (props: ExtendedActionSheetRootProps) => {
+export const ExtendedActionSheetRoot = (
+  props: ExtendedActionSheetRootProps,
+) => {
   const { children, ...otherProps } = props;
   return (
     <SeedExtendedActionSheet.Root closeOnInteractOutside={true} {...otherProps}>
@@ -17,10 +24,20 @@ export const ExtendedActionSheetRoot = (props: ExtendedActionSheetRootProps) => 
   );
 };
 
-export interface ExtendedActionSheetTriggerProps extends SeedExtendedActionSheet.TriggerProps {}
+/**
+ * @deprecated Use menu-sheet instead.
+ */
+export interface ExtendedActionSheetTriggerProps
+  extends SeedExtendedActionSheet.TriggerProps {}
 
+/**
+ * @deprecated Use menu-sheet instead.
+ */
 export const ExtendedActionSheetTrigger = SeedExtendedActionSheet.Trigger;
 
+/**
+ * @deprecated Use menu-sheet instead.
+ */
 export interface ExtendedActionSheetContentProps
   extends Omit<SeedExtendedActionSheet.ContentProps, "title"> {
   title?: React.ReactNode;
@@ -28,6 +45,9 @@ export interface ExtendedActionSheetContentProps
   layerIndex?: number;
 }
 
+/**
+ * @deprecated Use menu-sheet instead.
+ */
 export const ExtendedActionSheetContent = forwardRef<
   HTMLDivElement,
   ExtendedActionSheetContentProps
@@ -51,23 +71,46 @@ export const ExtendedActionSheetContent = forwardRef<
       <SeedExtendedActionSheet.Content ref={ref} {...otherProps}>
         {title && (
           <SeedExtendedActionSheet.Header>
-            <SeedExtendedActionSheet.Title>{title}</SeedExtendedActionSheet.Title>
+            <SeedExtendedActionSheet.Title>
+              {title}
+            </SeedExtendedActionSheet.Title>
           </SeedExtendedActionSheet.Header>
         )}
         <SeedExtendedActionSheet.List>{children}</SeedExtendedActionSheet.List>
         <SeedExtendedActionSheet.Footer>
           {/* You may implement your own i18n for dismiss label */}
-          <SeedExtendedActionSheet.CloseButton>취소</SeedExtendedActionSheet.CloseButton>
+          <SeedExtendedActionSheet.CloseButton>
+            취소
+          </SeedExtendedActionSheet.CloseButton>
         </SeedExtendedActionSheet.Footer>
       </SeedExtendedActionSheet.Content>
     </SeedExtendedActionSheet.Positioner>
   );
 });
 
-export interface ExtendedActionSheetGroupProps extends SeedExtendedActionSheet.GroupProps {}
+/**
+ * @deprecated Use menu-sheet instead.
+ */
+export interface ExtendedActionSheetGroupProps
+  extends SeedExtendedActionSheet.GroupProps {}
 
+/**
+ * @deprecated Use menu-sheet instead.
+ */
 export const ExtendedActionSheetGroup = SeedExtendedActionSheet.Group;
 
-export interface ExtendedActionSheetItemProps extends SeedExtendedActionSheet.ItemProps {}
+/**
+ * @deprecated Use menu-sheet instead.
+ */
+export interface ExtendedActionSheetItemProps
+  extends SeedExtendedActionSheet.ItemProps {}
 
+/**
+ * @deprecated Use menu-sheet instead.
+ */
 export const ExtendedActionSheetItem = SeedExtendedActionSheet.Item;
+
+/**
+ * This file is generated snippet from the Seed Design.
+ * You can extend the functionality from this snippet if needed.
+ */