+++ packages/qvism-preset/src/recipes.ts
@@ -8,6 +8,7 @@ import badge from "./recipes/badge";
 import bottomSheet from "./recipes/bottom-sheet";
 import callout from "./recipes/callout";
 import checkbox from "./recipes/checkbox";
+import chip from "./recipes/chip";
 import chipTabs from "./recipes/chip-tabs";
 import contextualFloatingButton from "./recipes/contextual-floating-button";
 import controlChip from "./recipes/control-chip";
@@ -23,6 +24,8 @@ import inlineBanner from "./recipes/inline-banner";
 import linkContent from "./recipes/link-content";
 import mannerTemp from "./recipes/manner-temp";
 import mannerTempBadge from "./recipes/manner-temp-badge";
+import menuSheet from "./recipes/menu-sheet";
+import menuSheetItem from "./recipes/menu-sheet-item";
 import { notificationBadge, notificationBadgePositioner } from "./recipes/notification-badge";
 import progressCircle from "./recipes/progress-circle";
 import pullToRefresh from "./recipes/pull-to-refresh";
@@ -60,10 +63,13 @@ export const recipes = {
   dialog,
   checkbox,
   actionChip,
+  chip,
   controlChip,
   callout,
   mannerTemp,
   mannerTempBadge,
+  menuSheet,
+  menuSheetItem,
   segmentedControl,
   selectBoxGroup,
   selectBox,