+++ packages/figma/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @seed-design/figma
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [4afe80b]
+  - @seed-design/css@0.1.7
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [235147d]
+- Updated dependencies [3c13ad7]
+  - @seed-design/css@0.1.6
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [861ecb4]
+- Updated dependencies [3889eb6]
+  - @seed-design/css@0.1.5
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [0ffcd48]
+  - @seed-design/css@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes