+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @seed-design/mcp
 
+## 0.1.7
+
+### Patch Changes
+
+- @seed-design/figma@0.1.7
+
+## 0.1.6
+
+### Patch Changes
+
+- @seed-design/figma@0.1.6
+
+## 0.1.5
+
+### Patch Changes
+
+- @seed-design/figma@0.1.5
+
+## 0.1.4
+
+### Patch Changes
+
+- @seed-design/figma@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes